refactor(components): refactored routes imports and exports
