+++ src/components/views/index.js
@@ -1,6 +1,5 @@
-// Barrel file for presentational components, which will be used to compose our smart containers
+// Barrel file for container components, which will be used to compose our main App component
 export { default as HeaderView } from './HeaderView';
-export { default as LandingView } from './modules/Main/Landing/LandingView';
+export { default as MainView } from './MainView';
 export { default as FooterView } from './FooterView';
-export { default as LoginFormView } from './modules/Header/LoginFormView';
-export { default as LogoutView } from './modules/Header/LogoutView';
+export { default as RoutesView } from './RoutesView';
